// packages/competencies/src/types/generated.d.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends {
    [key: string]: unknown;
}> = {
    [K in keyof T]: T[K];
};
// ...
export type Scalars = {
    ID: {
        input: string;
        output: string;
    };
    String: {
        input: string;
        output: string;
    };
    Boolean: {
        input: boolean;
        output: boolean;
    };
    Int: {
        input: number;
        output: number;
    };
    Float: {
        input: number;
        output: number;
    };
};
export type Competency = {
    __typename?: 'Competency';
    advantages?: Maybe<Array<Maybe<Scalars['String']['output']>>>;
    appelations?: Maybe<Array<Maybe<Scalars['String']['output']>>>;
    conclusion?: Maybe<Scalars['String']['output']>;
    definition?: Maybe<Scalars['String']['output']>;
    development?: Maybe<Scalars['String']['output']>;
    examples?: Maybe<Array<Maybe<Scalars['String']['output']>>>;
    extension?: Maybe<Scalars['String']['output']>;
    id?: Maybe<Scalars['ID']['output']>;
    importance?: Maybe<Scalars['String']['output']>;
    keywords?: Maybe<Array<Maybe<Scalars['String']['output']>>>;
    relatedSkills?: Maybe<Array<Maybe<Scalars['String']['output']>>>;
    slug?: Maybe<Scalars['String']['output']>;
    title?: Maybe<Scalars['String']['output']>;
};
export type CompetencyInput = {
    title?: InputMaybe<Scalars['String']['input']>;
};
export type PageParamsInput = {
    skip?: InputMaybe<Scalars['Int']['input']>;
    take?: InputMaybe<Scalars['Int']['input']>;
};
// ...
export type GetManyCompetenciesQueryVariables = Exact<{
    data?: InputMaybe<CompetencyInput>;
    params?: InputMaybe<PageParamsInput>;
    includeDefinition?: InputMaybe<Scalars['Boolean']['input']>;
}>;